Consolidate image type
