+++ src/data/Image.ts
@@ -1,0 +1,10 @@
+export interface Image {
+    copyright: string;
+    date: string;
+    explanation: string;
+    hdurl: string;
+    media_type: string;
+    service_version: string;
+    title: string;
+    url: string;
+}
